fix: keep planner map theme in sync
